+++ packages/eth-json-rpc-provider/src/provider-from-engine.test.ts
@@ -1,5 +1,5 @@
 import { JsonRpcEngine } from '@metamask/json-rpc-engine';
-import { promisify } from 'util';
+import { providerErrors } from '@metamask/rpc-errors';
 
 import { providerFromEngine } from './provider-from-engine';
 
@@ -11,33 +11,36 @@ describe('providerFromEngine', () => {
       end();
     });
     const provider = providerFromEngine(engine);
-    const promisifiedSendAsync = promisify(provider.sendAsync);
     const exampleRequest = {
       id: 1,
       jsonrpc: '2.0' as const,
       method: 'test',
     };
 
-    const response = await promisifiedSendAsync(exampleRequest);
+    const response = await provider.request(exampleRequest);
 
-    expect(response.result).toBe(42);
+    expect(response).toBe(42);
   });
 
   it('handle a failed request', async () => {
     const engine = new JsonRpcEngine();
-    engine.push((_req, _res, _next, _end) => {
-      throw new Error('Test error');
+    engine.push((_req, _res, _next, end) => {
+      end(
+        providerErrors.custom({
+          code: 1001,
+          message: 'Test error',
+        }),
+      );
     });
     const provider = providerFromEngine(engine);
-    const promisifiedSendAsync = promisify(provider.sendAsync);
     const exampleRequest = {
       id: 1,
       jsonrpc: '2.0' as const,
       method: 'test',
     };
 
-    await expect(async () =>
-      promisifiedSendAsync(exampleRequest),
-    ).rejects.toThrow('Test error');
+    await expect(async () => provider.request(exampleRequest)).rejects.toThrow(
+      'Test error',
+    );
   });
 });