+++ packages/eth-json-rpc-provider/src/provider-from-middleware.test.ts
@@ -1,5 +1,5 @@
 import type { JsonRpcMiddleware } from '@metamask/json-rpc-engine';
-import { promisify } from 'util';
+import { providerErrors } from '@metamask/rpc-errors';
 
 import { providerFromMiddleware } from './provider-from-middleware';
 
@@ -12,32 +12,34 @@ describe('providerFromMiddleware', () => {
       end();
     };
     const provider = providerFromMiddleware(middleware);
-    const promisifiedSendAsync = promisify(provider.sendAsync);
     const exampleRequest = {
       id: 1,
       jsonrpc: '2.0' as const,
       method: 'test',
     };
 
-    const response = await promisifiedSendAsync(exampleRequest);
+    const response = await provider.request(exampleRequest);
 
-    expect(response.result).toBe(42);
+    expect(response).toBe(42);
   });
 
   it('handle a failed request', async () => {
-    const middleware = () => {
-      throw new Error('Test error');
-    };
-    const provider = providerFromMiddleware(middleware);
-    const promisifiedSendAsync = promisify(provider.sendAsync);
+    const provider = providerFromMiddleware((_req, _res, _next, end) => {
+      end(
+        providerErrors.custom({
+          code: 1001,
+          message: 'Test error',
+        }),
+      );
+    });
     const exampleRequest = {
       id: 1,
       jsonrpc: '2.0' as const,
       method: 'test',
     };
 
-    await expect(async () =>
-      promisifiedSendAsync(exampleRequest),
-    ).rejects.toThrow('Test error');
+    await expect(async () => provider.request(exampleRequest)).rejects.toThrow(
+      'Test error',
+    );
   });
 });