+++ packages/gas-fee-controller/src/GasFeeController.test.ts
@@ -23,11 +23,7 @@ import {
   fetchEthGasPriceEstimate,
   calculateTimeEstimate,
 } from './gas-util';
-import {
-  GAS_API_BASE_URL,
-  GAS_ESTIMATE_TYPES,
-  GasFeeController,
-} from './GasFeeController';
+import { GAS_ESTIMATE_TYPES, GasFeeController } from './GasFeeController';
 import type {
   GasFeeState,
   GasFeeStateChange,
@@ -228,12 +224,13 @@ describe('GasFeeController', () => {
    * GasFeeController.
    * @param options.getCurrentNetworkLegacyGasAPICompatibility - Sets
    * getCurrentNetworkLegacyGasAPICompatibility on the GasFeeController.
+   * @param options.legacyAPIEndpoint - Sets legacyAPIEndpoint on the GasFeeController.
+   * @param options.EIP1559APIEndpoint - Sets EIP1559APIEndpoint on the GasFeeController.
    * @param options.clientId - Sets clientId on the GasFeeController.
    * @param options.networkControllerState - State object to initialize
    * NetworkController with.
    * @param options.interval - The polling interval.
    * @param options.state - The initial GasFeeController state
-   * @param options.infuraAPIKey - The Infura API key.
    * @param options.initializeNetworkProvider - Whether to instruct the
    * NetworkController to initialize its provider.
    */
@@ -242,7 +239,8 @@ describe('GasFeeController', () => {
     getCurrentNetworkLegacyGasAPICompatibility = jest
       .fn()
       .mockReturnValue(false),
-    infuraAPIKey = 'INFURA_API_KEY',
+    legacyAPIEndpoint = 'http://legacy.endpoint/<chain_id>',
+    EIP1559APIEndpoint = 'http://eip-1559.endpoint/<chain_id>',
     clientId,
     getChainId,
     onNetworkDidChange,
@@ -255,11 +253,13 @@ describe('GasFeeController', () => {
     onNetworkDidChange?: jest.Mock<void>;
     getIsEIP1559Compatible?: jest.Mock<Promise<boolean>>;
     getCurrentNetworkLegacyGasAPICompatibility?: jest.Mock<boolean>;
+    legacyAPIEndpoint?: string;
+    // eslint-disable-next-line @typescript-eslint/naming-convention
+    EIP1559APIEndpoint?: string;
     clientId?: string;
     networkControllerState?: Partial<NetworkState>;
     state?: GasFeeState;
     interval?: number;
-    infuraAPIKey?: string;
     initializeNetworkProvider?: boolean;
   } = {}) {
     const controllerMessenger = getControllerMessenger();
@@ -277,10 +277,11 @@ describe('GasFeeController', () => {
       messenger,
       getCurrentNetworkLegacyGasAPICompatibility,
       getCurrentNetworkEIP1559Compatibility: getIsEIP1559Compatible, // change this for networkDetails.state.networkDetails.isEIP1559Compatible ???
+      legacyAPIEndpoint,
+      EIP1559APIEndpoint,
       state,
       clientId,
       interval,
-      infuraAPIKey,
     });
   }
 
@@ -298,7 +299,6 @@ describe('GasFeeController', () => {
     // eslint-disable-next-line @typescript-eslint/no-floating-promises
     blockTracker?.destroy();
     sinon.restore();
-    jest.clearAllMocks();
   });
 
   describe('constructor', () => {
@@ -335,6 +335,8 @@ describe('GasFeeController', () => {
             getCurrentNetworkLegacyGasAPICompatibility: jest
               .fn()
               .mockReturnValue(true),
+            legacyAPIEndpoint: 'https://some-legacy-endpoint/<chain_id>',
+            EIP1559APIEndpoint: 'https://some-eip-1559-endpoint/<chain_id>',
             networkControllerState: {
               networkConfigurations: {
                 'AAAA-BBBB-CCCC-DDDD': {
@@ -362,14 +364,14 @@ describe('GasFeeController', () => {
             isEIP1559Compatible: false,
             isLegacyGasAPICompatible: true,
             fetchGasEstimates,
-            fetchGasEstimatesUrl: `${GAS_API_BASE_URL}/networks/1337/suggestedGasFees`,
+            fetchGasEstimatesUrl: 'https://some-eip-1559-endpoint/1337',
             fetchLegacyGasPriceEstimates,
-            fetchLegacyGasPriceEstimatesUrl: `${GAS_API_BASE_URL}/networks/1337/gasPrices`,
+            fetchLegacyGasPriceEstimatesUrl:
+              'https://some-legacy-endpoint/1337',
             fetchEthGasPriceEstimate,
             calculateTimeEstimate,
             clientId: '99999',
             ethQuery: expect.any(EthQuery),
-            infuraAPIKey: expect.any(String),
             nonRPCGasFeeApisDisabled: false,
           });
         });
@@ -399,6 +401,8 @@ describe('GasFeeController', () => {
             getCurrentNetworkLegacyGasAPICompatibility: jest
               .fn()
               .mockReturnValue(true),
+            legacyAPIEndpoint: 'https://some-legacy-endpoint/<chain_id>',
+            EIP1559APIEndpoint: 'https://some-eip-1559-endpoint/<chain_id>',
             networkControllerState: {
               networkConfigurations: {
                 'AAAA-BBBB-CCCC-DDDD': {
@@ -428,14 +432,14 @@ describe('GasFeeController', () => {
             isEIP1559Compatible: false,
             isLegacyGasAPICompatible: true,
             fetchGasEstimates,
-            fetchGasEstimatesUrl: `${GAS_API_BASE_URL}/networks/1337/suggestedGasFees`,
+            fetchGasEstimatesUrl: 'https://some-eip-1559-endpoint/1337',
             fetchLegacyGasPriceEstimates,
-            fetchLegacyGasPriceEstimatesUrl: `${GAS_API_BASE_URL}/networks/1337/gasPrices`,
+            fetchLegacyGasPriceEstimatesUrl:
+              'https://some-legacy-endpoint/1337',
             fetchEthGasPriceEstimate,
             calculateTimeEstimate,
             clientId: '99999',
             ethQuery: expect.any(EthQuery),
-            infuraAPIKey: expect.any(String),
             nonRPCGasFeeApisDisabled: false,
           });
         });
@@ -735,12 +739,12 @@ describe('GasFeeController', () => {
 
   describe('fetchGasFeeEstimates', () => {
     describe('when on any network supporting legacy gas estimation api', () => {
-      const defaultConstructorOptions = {
+      const getDefaultOptions = () => ({
         getIsEIP1559Compatible: jest.fn().mockResolvedValue(false),
         getCurrentNetworkLegacyGasAPICompatibility: jest
           .fn()
           .mockReturnValue(true),
-      };
+      });
       const mockDetermineGasFeeCalculations = buildMockGasFeeStateLegacy();
 
       beforeEach(() => {
@@ -751,7 +755,9 @@ describe('GasFeeController', () => {
 
       it('should call determineGasFeeCalculations correctly', async () => {
         await setupGasFeeController({
-          ...defaultConstructorOptions,
+          ...getDefaultOptions(),
+          legacyAPIEndpoint: 'https://some-legacy-endpoint/<chain_id>',
+          EIP1559APIEndpoint: 'https://some-eip-1559-endpoint/<chain_id>',
           networkControllerState: {
             networkConfigurations: {
               'AAAA-BBBB-CCCC-DDDD': {
@@ -779,20 +785,19 @@ describe('GasFeeController', () => {
           isEIP1559Compatible: false,
           isLegacyGasAPICompatible: true,
           fetchGasEstimates,
-          fetchGasEstimatesUrl: `${GAS_API_BASE_URL}/networks/1337/suggestedGasFees`,
+          fetchGasEstimatesUrl: 'https://some-eip-1559-endpoint/1337',
           fetchLegacyGasPriceEstimates,
-          fetchLegacyGasPriceEstimatesUrl: `${GAS_API_BASE_URL}/networks/1337/gasPrices`,
+          fetchLegacyGasPriceEstimatesUrl: 'https://some-legacy-endpoint/1337',
           fetchEthGasPriceEstimate,
           calculateTimeEstimate,
           clientId: '99999',
           ethQuery: expect.any(EthQuery),
-          infuraAPIKey: expect.any(String),
           nonRPCGasFeeApisDisabled: false,
         });
       });
 
       it('should update the state with a fetched set of estimates', async () => {
-        await setupGasFeeController(defaultConstructorOptions);
+        await setupGasFeeController(getDefaultOptions());
 
         await gasFeeController.fetchGasFeeEstimates();
 
@@ -802,7 +807,7 @@ describe('GasFeeController', () => {
       });
 
       it('should return the same data that it puts into state', async () => {
-        await setupGasFeeController(defaultConstructorOptions);
+        await setupGasFeeController(getDefaultOptions());
 
         const estimateData = await gasFeeController.fetchGasFeeEstimates();
 
@@ -811,7 +816,8 @@ describe('GasFeeController', () => {
 
       it('should call determineGasFeeCalculations correctly when getChainId returns a number input', async () => {
         await setupGasFeeController({
-          ...defaultConstructorOptions,
+          ...getDefaultOptions(),
+          legacyAPIEndpoint: 'http://legacy.endpoint/<chain_id>',
           getChainId: jest.fn().mockReturnValue(1),
         });
 
@@ -819,14 +825,15 @@ describe('GasFeeController', () => {
 
         expect(mockedDetermineGasFeeCalculations).toHaveBeenCalledWith(
           expect.objectContaining({
-            fetchLegacyGasPriceEstimatesUrl: `${GAS_API_BASE_URL}/networks/1/gasPrices`,
+            fetchLegacyGasPriceEstimatesUrl: 'http://legacy.endpoint/1',
           }),
         );
       });
 
       it('should call determineGasFeeCalculations correctly when getChainId returns a hexstring input', async () => {
         await setupGasFeeController({
-          ...defaultConstructorOptions,
+          ...getDefaultOptions(),
+          legacyAPIEndpoint: 'http://legacy.endpoint/<chain_id>',
           getChainId: jest.fn().mockReturnValue('0x1'),
         });
 
@@ -834,14 +841,14 @@ describe('GasFeeController', () => {
 
         expect(mockedDetermineGasFeeCalculations).toHaveBeenCalledWith(
           expect.objectContaining({
-            fetchLegacyGasPriceEstimatesUrl: `${GAS_API_BASE_URL}/networks/1/gasPrices`,
+            fetchLegacyGasPriceEstimatesUrl: 'http://legacy.endpoint/1',
           }),
         );
       });
 
       it('should call determineGasFeeCalculations correctly when nonRPCGasFeeApisDisabled is true', async () => {
         await setupGasFeeController({
-          ...defaultConstructorOptions,
+          ...getDefaultOptions(),
           state: {
             ...buildMockGasFeeStateEthGasPrice(),
             nonRPCGasFeeApisDisabled: true,
@@ -859,7 +866,7 @@ describe('GasFeeController', () => {
 
       it('should call determineGasFeeCalculations correctly when nonRPCGasFeeApisDisabled is false', async () => {
         await setupGasFeeController({
-          ...defaultConstructorOptions,
+          ...getDefaultOptions(),
           state: {
             ...buildMockGasFeeStateEthGasPrice(),
             nonRPCGasFeeApisDisabled: false,
@@ -877,7 +884,8 @@ describe('GasFeeController', () => {
 
       it('should call determineGasFeeCalculations correctly when getChainId returns a numeric string input', async () => {
         await setupGasFeeController({
-          ...defaultConstructorOptions,
+          ...getDefaultOptions(),
+          legacyAPIEndpoint: 'http://legacy.endpoint/<chain_id>',
           getChainId: jest.fn().mockReturnValue('1'),
         });
 
@@ -885,16 +893,16 @@ describe('GasFeeController', () => {
 
         expect(mockedDetermineGasFeeCalculations).toHaveBeenCalledWith(
           expect.objectContaining({
-            fetchLegacyGasPriceEstimatesUrl: `${GAS_API_BASE_URL}/networks/1/gasPrices`,
+            fetchLegacyGasPriceEstimatesUrl: 'http://legacy.endpoint/1',
           }),
         );
       });
     });
 
     describe('when on any network supporting EIP-1559', () => {
-      const defaultConstructorOptions = {
+      const getDefaultOptions = () => ({
         getIsEIP1559Compatible: jest.fn().mockResolvedValue(true),
-      };
+      });
       const mockDetermineGasFeeCalculations = buildMockGasFeeStateFeeMarket();
 
       beforeEach(() => {
@@ -905,7 +913,9 @@ describe('GasFeeController', () => {
 
       it('should call determineGasFeeCalculations correctly', async () => {
         await setupGasFeeController({
-          ...defaultConstructorOptions,
+          ...getDefaultOptions(),
+          legacyAPIEndpoint: 'https://some-legacy-endpoint/<chain_id>',
+          EIP1559APIEndpoint: 'https://some-eip-1559-endpoint/<chain_id>',
           networkControllerState: {
             networkConfigurations: {
               'AAAA-BBBB-CCCC-DDDD': {
@@ -933,20 +943,19 @@ describe('GasFeeController', () => {
           isEIP1559Compatible: true,
           isLegacyGasAPICompatible: false,
           fetchGasEstimates,
-          fetchGasEstimatesUrl: `${GAS_API_BASE_URL}/networks/1337/suggestedGasFees`,
+          fetchGasEstimatesUrl: 'https://some-eip-1559-endpoint/1337',
           fetchLegacyGasPriceEstimates,
-          fetchLegacyGasPriceEstimatesUrl: `${GAS_API_BASE_URL}/networks/1337/gasPrices`,
+          fetchLegacyGasPriceEstimatesUrl: 'https://some-legacy-endpoint/1337',
           fetchEthGasPriceEstimate,
           calculateTimeEstimate,
           clientId: '99999',
           ethQuery: expect.any(EthQuery),
-          infuraAPIKey: expect.any(String),
           nonRPCGasFeeApisDisabled: false,
         });
       });
 
       it('should update the state with a fetched set of estimates', async () => {
-        await setupGasFeeController(defaultConstructorOptions);
+        await setupGasFeeController(getDefaultOptions());
 
         await gasFeeController.fetchGasFeeEstimates();
 
@@ -956,7 +965,7 @@ describe('GasFeeController', () => {
       });
 
       it('should return the same data that it puts into state', async () => {
-        await setupGasFeeController(defaultConstructorOptions);
+        await setupGasFeeController(getDefaultOptions());
 
         const estimateData = await gasFeeController.fetchGasFeeEstimates();
 
@@ -965,7 +974,8 @@ describe('GasFeeController', () => {
 
       it('should call determineGasFeeCalculations with a URL that contains the chain ID', async () => {
         await setupGasFeeController({
-          ...defaultConstructorOptions,
+          ...getDefaultOptions(),
+          EIP1559APIEndpoint: 'http://eip-1559.endpoint/<chain_id>',
           getChainId: jest.fn().mockReturnValue('0x1'),
         });
 
@@ -973,13 +983,13 @@ describe('GasFeeController', () => {
 
         expect(mockedDetermineGasFeeCalculations).toHaveBeenCalledWith(
           expect.objectContaining({
-            fetchGasEstimatesUrl: `${GAS_API_BASE_URL}/networks/1/suggestedGasFees`,
+            fetchGasEstimatesUrl: 'http://eip-1559.endpoint/1',
           }),
         );
       });
     });
     describe('when passed a networkClientId in options object', () => {
-      const defaultConstructorOptions = {
+      const getDefaultOptions = () => ({
         getIsEIP1559Compatible: jest.fn().mockResolvedValue(true),
         networkControllerState: {
           networksMetadata: {
@@ -1003,7 +1013,7 @@ describe('GasFeeController', () => {
             },
           },
         },
-      };
+      });
       const mockDetermineGasFeeCalculations = buildMockGasFeeStateFeeMarket();
 
       beforeEach(() => {
@@ -1014,7 +1024,9 @@ describe('GasFeeController', () => {
 
       it('should call determineGasFeeCalculations correctly', async () => {
         await setupGasFeeController({
-          ...defaultConstructorOptions,
+          ...getDefaultOptions(),
+          legacyAPIEndpoint: 'https://some-legacy-endpoint/<chain_id>',
+          EIP1559APIEndpoint: 'https://some-eip-1559-endpoint/<chain_id>',
           clientId: '99999',
         });
 
@@ -1028,20 +1040,17 @@ describe('GasFeeController', () => {
           fetchGasEstimates,
           // TODO: Either fix this lint violation or explain why it's necessary to ignore.
           // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
-          fetchGasEstimatesUrl: `${GAS_API_BASE_URL}/networks/${convertHexToDecimal(
-            ChainId.goerli,
-          )}/suggestedGasFees`,
+          fetchGasEstimatesUrl: 'https://some-eip-1559-endpoint/5',
           fetchLegacyGasPriceEstimates,
           // TODO: Either fix this lint violation or explain why it's necessary to ignore.
           // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
-          fetchLegacyGasPriceEstimatesUrl: `${GAS_API_BASE_URL}/networks/${convertHexToDecimal(
+          fetchLegacyGasPriceEstimatesUrl: `https://some-legacy-endpoint/${convertHexToDecimal(
             ChainId.goerli,
-          )}/gasPrices`,
+          )}`,
           fetchEthGasPriceEstimate,
           calculateTimeEstimate,
           clientId: '99999',
           ethQuery: expect.any(EthQuery),
-          infuraAPIKey: expect.any(String),
           nonRPCGasFeeApisDisabled: false,
         });
       });
@@ -1049,7 +1058,7 @@ describe('GasFeeController', () => {
       describe("the chainId of the networkClientId matches the globally selected network's chainId", () => {
         it('should update the globally selected network state with a fetched set of estimates', async () => {
           await setupGasFeeController({
-            ...defaultConstructorOptions,
+            ...getDefaultOptions(),
             getChainId: jest.fn().mockReturnValue(ChainId.goerli),
             onNetworkDidChange: jest.fn(),
           });
@@ -1065,7 +1074,7 @@ describe('GasFeeController', () => {
 
         it('should update the gasFeeEstimatesByChainId state with a fetched set of estimates', async () => {
           await setupGasFeeController({
-            ...defaultConstructorOptions,
+            ...getDefaultOptions(),
             getChainId: jest.fn().mockReturnValue(ChainId.goerli),
             onNetworkDidChange: jest.fn(),
           });
@@ -1083,7 +1092,7 @@ describe('GasFeeController', () => {
       describe("the chainId of the networkClientId does not match the globally selected network's chainId", () => {
         it('should not update the globally selected network state with a fetched set of estimates', async () => {
           await setupGasFeeController({
-            ...defaultConstructorOptions,
+            ...getDefaultOptions(),
             getChainId: jest.fn().mockReturnValue(ChainId.mainnet),
             onNetworkDidChange: jest.fn(),
           });
@@ -1101,7 +1110,7 @@ describe('GasFeeController', () => {
 
         it('should update the gasFeeEstimatesByChainId state with a fetched set of estimates', async () => {
           await setupGasFeeController({
-            ...defaultConstructorOptions,
+            ...getDefaultOptions(),
             getChainId: jest.fn().mockReturnValue(ChainId.mainnet),
             onNetworkDidChange: jest.fn(),
           });
@@ -1117,7 +1126,7 @@ describe('GasFeeController', () => {
       });
 
       it('should return the same data that it puts into state', async () => {
-        await setupGasFeeController(defaultConstructorOptions);
+        await setupGasFeeController(getDefaultOptions());
 
         const estimateData = await gasFeeController.fetchGasFeeEstimates({
           networkClientId: 'sepolia',
@@ -1128,7 +1137,8 @@ describe('GasFeeController', () => {
 
       it('should call determineGasFeeCalculations with a URL that contains the chain ID', async () => {
         await setupGasFeeController({
-          ...defaultConstructorOptions,
+          ...getDefaultOptions(),
+          EIP1559APIEndpoint: 'http://eip-1559.endpoint/<chain_id>',
         });
 
         await gasFeeController.fetchGasFeeEstimates({
@@ -1139,9 +1149,9 @@ describe('GasFeeController', () => {
           expect.objectContaining({
             // TODO: Either fix this lint violation or explain why it's necessary to ignore.
             // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
-            fetchGasEstimatesUrl: `${GAS_API_BASE_URL}/networks/${convertHexToDecimal(
+            fetchGasEstimatesUrl: `http://eip-1559.endpoint/${convertHexToDecimal(
               ChainId.sepolia,
-            )}/suggestedGasFees`,
+            )}`,
           }),
         );
       });
@@ -1156,6 +1166,8 @@ describe('GasFeeController', () => {
         getCurrentNetworkLegacyGasAPICompatibility: jest
           .fn()
           .mockReturnValue(true),
+        legacyAPIEndpoint: 'https://some-legacy-endpoint/<chain_id>',
+        EIP1559APIEndpoint: 'https://some-eip-1559-endpoint/<chain_id>',
         networkControllerState: {
           networksMetadata: {
             goerli: {
@@ -1183,9 +1195,9 @@ describe('GasFeeController', () => {
         expect.objectContaining({
           // TODO: Either fix this lint violation or explain why it's necessary to ignore.
           // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
-          fetchGasEstimatesUrl: `${GAS_API_BASE_URL}/networks/${convertHexToDecimal(
+          fetchGasEstimatesUrl: `https://some-eip-1559-endpoint/${convertHexToDecimal(
             ChainId.goerli,
-          )}/suggestedGasFees`,
+          )}`,
         }),
       );
       await clock.tickAsync(pollingInterval / 2);
@@ -1196,9 +1208,9 @@ describe('GasFeeController', () => {
         expect.objectContaining({
           // TODO: Either fix this lint violation or explain why it's necessary to ignore.
           // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
-          fetchGasEstimatesUrl: `${GAS_API_BASE_URL}/networks/${convertHexToDecimal(
+          fetchGasEstimatesUrl: `https://some-eip-1559-endpoint/${convertHexToDecimal(
             ChainId.goerli,
-          )}/suggestedGasFees`,
+          )}`,
         }),
       );
       expect(
@@ -1211,9 +1223,9 @@ describe('GasFeeController', () => {
         expect.objectContaining({
           // TODO: Either fix this lint violation or explain why it's necessary to ignore.
           // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
-          fetchGasEstimatesUrl: `${GAS_API_BASE_URL}/networks/${convertHexToDecimal(
+          fetchGasEstimatesUrl: `https://some-eip-1559-endpoint/${convertHexToDecimal(
             ChainId.sepolia,
-          )}/suggestedGasFees`,
+          )}`,
         }),
       );
     });