+++ packages/profile-sync-controller/src/controllers/authentication/AuthenticationController.ts
@@ -37,7 +37,7 @@ type SessionData = {
 };
 
 type MetaMetricsAuth = {
-  getMetaMetricsId: () => string;
+  getMetaMetricsId: () => string | Promise<string>;
   agent: 'extension' | 'mobile';
 };
 
@@ -49,7 +49,9 @@ export type AuthenticationControllerState = {
   isSignedIn: boolean;
   sessionData?: SessionData;
 };
-const defaultState: AuthenticationControllerState = { isSignedIn: false };
+export const defaultState: AuthenticationControllerState = {
+  isSignedIn: false,
+};
 const metadata: StateMetadata<AuthenticationControllerState> = {
   isSignedIn: {
     persist: true,
@@ -129,6 +131,10 @@ export default class AuthenticationController extends BaseController<
       name: controllerName,
       state: { ...defaultState, ...state },
     });
+
+    if (!metametrics) {
+      throw new Error('`metametrics` field is required');
+    }
 
     this.#metametrics = metametrics;
 
@@ -236,7 +242,7 @@ export default class AuthenticationController extends BaseController<
       const rawMessage = createLoginRawMessage(nonce, publicKey);
       const signature = await this.#snapSignMessage(rawMessage);
       const loginResponse = await login(rawMessage, signature, {
-        metametricsId: this.#metametrics.getMetaMetricsId(),
+        metametricsId: await this.#metametrics.getMetaMetricsId(),
         agent: this.#metametrics.agent,
       });
       if (!loginResponse?.token) {
@@ -249,7 +255,10 @@ export default class AuthenticationController extends BaseController<
       };
 
       // 3. Trade for Access Token
-      const accessToken = await getAccessToken(loginResponse.token);
+      const accessToken = await getAccessToken(
+        loginResponse.token,
+        this.#metametrics.agent,
+      );
       if (!accessToken) {
         throw new Error(`Unable to get Access Token`);
       }